requirements.txt
absl-py==2.1.0
altair==5.2.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
comtypes==1.4.6
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1rc1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.49.0
fqdn==1.5.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
grpcio==1.65.5
gTTS==2.5.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.5.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
openai==1.41.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==24.1
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pillow==10.4.0
platformdirs==4.2.0
pooch==1.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.4
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.8.1b0
Pygments==2.18.0
pyparsing==3.1.2
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pyttsx3==2.91
pytz==2024.1
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.33.0
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.4
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==72.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.4.0
SpeechRecognition==3.10.4
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.37.0
sympy==1.13.2
tenacity==8.2.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.17.0
threadpoolctl==3.3.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torc==0.1.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
transformers==4.44.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
watchdog==4.0.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.44.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xgboost==2.0.3
